# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ohana API",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/api/health": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "status",
                    "checks"
                  ],
                  "properties": {
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "ok"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "degraded"
                          ]
                        }
                      ]
                    },
                    "checks": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "database",
                        "storage"
                      ],
                      "properties": {
                        "database": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "up"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "down"
                              ]
                            }
                          ]
                        },
                        "storage": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "up"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "down"
                              ]
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "status",
                    "checks"
                  ],
                  "properties": {
                    "status": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "ok"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "degraded"
                          ]
                        }
                      ]
                    },
                    "checks": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "database",
                        "storage"
                      ],
                      "properties": {
                        "database": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "up"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "down"
                              ]
                            }
                          ]
                        },
                        "storage": {
                          "anyOf": [
                            {
                              "type": "string",
                              "enum": [
                                "up"
                              ]
                            },
                            {
                              "type": "string",
                              "enum": [
                                "down"
                              ]
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces": {
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "additionalProperties": false,
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "timezone",
                      "revision",
                      "createdAt",
                      "updatedAt",
                      "memberCount"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "timezone": {
                        "minLength": 1,
                        "maxLength": 64,
                        "type": "string"
                      },
                      "revision": {
                        "pattern": "^[0-9]+$",
                        "type": "string"
                      },
                      "createdAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "memberCount": {
                        "minimum": 0,
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 200,
                    "type": "string"
                  },
                  "timezone": {
                    "minLength": 1,
                    "maxLength": 64,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "header",
            "name": "x-ohana-admin",
            "required": false
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "timezone",
                    "revision",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "timezone": {
                      "minLength": 1,
                      "maxLength": 64,
                      "type": "string"
                    },
                    "revision": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{spaceId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "spaceId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "timezone",
                    "revision",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "timezone": {
                      "minLength": 1,
                      "maxLength": 64,
                      "type": "string"
                    },
                    "revision": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 200,
                    "type": "string"
                  },
                  "timezone": {
                    "minLength": 1,
                    "maxLength": 64,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "spaceId",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "header",
            "name": "x-ohana-admin",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "id",
                    "name",
                    "timezone",
                    "revision",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "timezone": {
                      "minLength": 1,
                      "maxLength": 64,
                      "type": "string"
                    },
                    "revision": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{spaceId}/members": {
      "get": {
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "spaceId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "additionalProperties": false,
                    "type": "object",
                    "required": [
                      "id",
                      "spaceId",
                      "name",
                      "role",
                      "revision",
                      "createdAt",
                      "updatedAt"
                    ],
                    "properties": {
                      "id": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "spaceId": {
                        "format": "uuid",
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "displayName": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "phone": {
                        "type": "string"
                      },
                      "interfaceLanguage": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "ru"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "en"
                            ]
                          }
                        ]
                      },
                      "role": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "owner"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "regular"
                            ]
                          }
                        ]
                      },
                      "revision": {
                        "pattern": "^[0-9]+$",
                        "type": "string"
                      },
                      "createdAt": {
                        "format": "date-time",
                        "type": "string"
                      },
                      "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "name",
                  "role"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 200,
                    "type": "string"
                  },
                  "role": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "owner"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "regular"
                        ]
                      }
                    ]
                  },
                  "displayName": {
                    "minLength": 1,
                    "maxLength": 200,
                    "type": "string"
                  },
                  "email": {
                    "minLength": 3,
                    "maxLength": 200,
                    "type": "string"
                  },
                  "phone": {
                    "minLength": 3,
                    "maxLength": 40,
                    "type": "string"
                  },
                  "interfaceLanguage": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "ru"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "en"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "spaceId",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "header",
            "name": "x-ohana-admin",
            "required": false
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "id",
                    "spaceId",
                    "name",
                    "role",
                    "revision",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "spaceId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "interfaceLanguage": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "ru"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "en"
                          ]
                        }
                      ]
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "owner"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "regular"
                          ]
                        }
                      ]
                    },
                    "revision": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/spaces/{spaceId}/members/{memberId}": {
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "role"
                ],
                "properties": {
                  "role": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "owner"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "regular"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "spaceId",
            "required": true
          },
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "memberId",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "header",
            "name": "x-ohana-admin",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "id",
                    "spaceId",
                    "name",
                    "role",
                    "revision",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "spaceId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "interfaceLanguage": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "ru"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "en"
                          ]
                        }
                      ]
                    },
                    "role": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "owner"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "regular"
                          ]
                        }
                      ]
                    },
                    "revision": {
                      "pattern": "^[0-9]+$",
                      "type": "string"
                    },
                    "createdAt": {
                      "format": "date-time",
                      "type": "string"
                    },
                    "updatedAt": {
                      "format": "date-time",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/session": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "minLength": 1,
                    "maxLength": 1024,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "header",
            "name": "x-ohana-admin",
            "required": false
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      },
      "get": {
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "header",
            "name": "x-ohana-admin",
            "required": false
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/admin/password": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "currentPassword",
                  "newPassword"
                ],
                "properties": {
                  "currentPassword": {
                    "minLength": 1,
                    "maxLength": 1024,
                    "type": "string"
                  },
                  "newPassword": {
                    "minLength": 10,
                    "maxLength": 1024,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "header",
            "name": "x-ohana-admin",
            "required": false
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    }
  }
}